Fix Fog of War display (#2960)
* Fix player duplicated on map

* Fix ShowGameAsPlayer command

* Don't use `Player` as a map key anymore

After the change to have Player inherit PlayerProperties, any maps which
used the Player object as a key began misbehaving. This almost exclusively
affected fog of war, as we were storing a player's fog information in such
a Map. I've changed the affected Maps to use the player's userId as the
key instead.