Update ds.json
I realized the issue with my code generation function in google sheets. A single cell has a limited number of characters it can contain so it was only outputting up to the maximum number of characters it could contain, resulting in a much smaller Combined code output. I have manually added everything, so please let me know if this works.