Refactor agenda resolvers with ForAgainst abstraction (#3492)
* Refactor agenda resolvers with ForAgainst abstraction

* Simplify java.util references using var

* F

* f

* f

* f

* Refactor RevealPublicObjectiveService channel handling

Refactored methods in RevealPublicObjectiveService to obtain the MessageChannel from the event instead of passing it as a parameter. Updated all usages and related handlers to match the new method signatures, simplifying the API and reducing redundant parameter passing.

* Update RevealStage2.java

* Update AutoCompleteProvider.java

* Update AutoCompleteProvider.java

* Refactor ChecksAndBalances to use shared SC button logic

Removed the local getPlayerOptionsForChecksNBalances method and now use PickStrategyCardButtonHandler's implementation instead. This reduces code duplication and centralizes the logic for generating strategy card pick buttons.

* F

* Update CombatRollService.java