Refactor and expand action deck 2 cards and data (#3970)
* Refactor and expand action deck 2 cards and data

Replaced 'classified_weapons_acd2' and 'vulnerability' cards with new 'classified_weapons1-4' cards, updated their references in deck files, and moved deprecated cards to legacy_data.json. Added new cards 'incompetence' and 'oracle', updated the 'corruption' and 'ubiquity' cards, and made minor text corrections. Also removed 'ambassadors' from the main deck and moved it to legacy data, and ensured all deck references are consistent with the new card aliases.

* Update 'Incompetence' action card text and window

Revised the 'window' and 'text' fields for the 'Incompetence' action card to clarify timing and effect. The new text specifies that hits are assigned to ships and clarifies when the card is played.

* Update acd2.json